Moves 2 large sample paywalls to separate files.
